Started moving rendering code for tooltips into a component agnostic directory
