+++ assets/src/scripts/components/dailyValueHydrograph/selectors/time-series-data.js
@@ -61,3 +61,5 @@ export const getCurrentTimeSeriesLineSegments = createSelector(
         return lineSegments;
     }
 );
+
+//export const getDataAtCursorOffset
