+++ assets/src/scripts/components/dailyValueHydrograph/selectors/time-series-data.spec.js
@@ -1,4 +1,4 @@
-import {getCurrentTimeSeriesLineSegments} from './time-series-lines';
+import {getCurrentTimeSeriesLineSegments} from './time-series-data';
 
 describe('time-series-line module', () => {
     describe('getCurrentTimeSeriesLineSegments', () => {
