+++ assets/src/scripts/store/observations-state-reducer.js
@@ -9,6 +9,14 @@ const setCurrentTimeSeriesId = function(observationsState, action) {
     };
 };
 
+const setDailyValueCursorOffset = function(observationsState, action) {
+    return {
+        ...observationsState,
+        cursorOffset: action.cursorOffset
+
+    };
+};
+
 /*
  * Slice reducer
  */
@@ -16,6 +24,8 @@ export const observationsStateReducer = function(observationsState={}, action) {
     switch (action.type) {
         case 'SET_CURRENT_TIME_SERIES_ID':
             return setCurrentTimeSeriesId(observationsState, action);
+        case 'SET_DAILY_VALUE_CURSOR_OFFSET':
+            return setDailyValueCursorOffset(observationsState, action);
         default:
             return observationsState;
     }
