+++ assets/src/scripts/store/observations-state-reducer.spec.js
@@ -11,4 +11,15 @@ describe('observationsStateReducer', () => {
             currentTimeSeriesId: '11111'
         });
     });
+
+    it('should handle SET_DAILY_VALUE_CURSOR_OFFSET', () => {
+        expect(observationsStateReducer({
+            cursorOffset: '12345'
+        }, {
+            type: 'SET_DAILY_VALUE_CURSOR_OFFSET',
+            cursorOffset: '55555'
+        })).toEqual({
+            cursorOffset: '55555'
+        });
+    });
 });
